refactor(core): fix circular dependencies by exporting members with modules
