--- packages/botonic-core/src/constants.ts
@@ -1,15 +1,0 @@
-export const Providers = Object.freeze({
-  Messaging: {
-    FACEBOOK: 'facebook',
-    GENERIC: 'generic',
-    IMBEE: 'imbee',
-    INTERCOM: 'intercom',
-    SMOOCH_WEB: 'smooch_web',
-    SMOOCH: 'smooch',
-    TELEGRAM: 'telegram',
-    TWITTER: 'twitter',
-    WEBCHAT: 'webchat',
-    WECHAT: 'wechat',
-    WHATSAPP: 'whatsapp',
-  },
-})
