+++ packages/botonic-core/src/models/events/base-event.ts
@@ -1,0 +1,14 @@
+export enum EventTypes {
+  CONNECTION = 'connection',
+  MESSAGE = 'message',
+  ACK = 'ack',
+  TRACK = 'track',
+}
+
+export interface BaseEvent {
+  eventId: string
+  userId: string
+  eventType: EventTypes
+  createdAt: string
+  modifiedAt?: string
+}
