+++ packages/botonic-core/src/models/events/botonic-event.ts
@@ -1,0 +1,24 @@
+import { ConnectionEvent } from './connections'
+import {
+  AudioMessageEvent,
+  CarouselMessageEvent,
+  CustomMessageEvent,
+  DocumentMessageEvent,
+  ImageMessageEvent,
+  LocationMessageEvent,
+  PostbackMessageEvent,
+  TextMessageEvent,
+  VideoMessageEvent,
+} from './message'
+
+export type BotonicEvent =
+  | TextMessageEvent
+  | PostbackMessageEvent
+  | AudioMessageEvent
+  | DocumentMessageEvent
+  | ImageMessageEvent
+  | VideoMessageEvent
+  | LocationMessageEvent
+  | CarouselMessageEvent
+  | CustomMessageEvent
+  | ConnectionEvent
