+++ packages/botonic-core/src/models/events/connections/connection-event.ts
@@ -1,0 +1,11 @@
+import { BaseEvent, EventTypes } from '../base-event'
+
+export enum ConnectionEventStatuses {
+  CONNECTED = 'connected',
+  DISCONNECTED = 'disconnected',
+}
+
+export interface ConnectionEvent extends BaseEvent {
+  eventType: EventTypes.CONNECTION
+  status: ConnectionEventStatuses
+}
