+++ packages/botonic-core/src/models/events/connections/index.ts
@@ -1,11 +1,1 @@
-import { BaseEvent, EventTypes } from '..'
-
-export enum ConnectionEventStatuses {
-  CONNECTED = 'connected',
-  DISCONNECTED = 'disconnected',
-}
-
-export interface ConnectionEvent extends BaseEvent {
-  eventType: EventTypes.CONNECTION
-  status: ConnectionEventStatuses
-}
+export * from './connection-event'
