+++ packages/botonic-core/src/models/events/message/index.ts
@@ -1,37 +1,9 @@
-import { BaseEvent } from '..'
-
-export enum MessageEventTypes {
-  AUDIO = 'audio',
-  CAROUSEL = 'carousel',
-  CUSTOM = 'custom',
-  DOCUMENT = 'document',
-  IMAGE = 'image',
-  LOCATION = 'location',
-  TEXT = 'text',
-  POSTBACK = 'postback',
-  VIDEO = 'video',
-  /**
-   * TODO: contact, buttonmessage?, webchatsettings, whatsapp template
-   */
-}
-
-export enum MessageEventAck {
-  DRAFT = 'draft',
-  READ = 'read',
-  RECEIVED = 'received',
-  SENT = 'sent',
-}
-
-export enum MessageEventFrom {
-  AGENT = 'agent',
-  BOT = 'bot',
-  USER = 'user',
-}
-
-export interface BotonicMessageEvent extends BaseEvent {
-  ack: MessageEventAck
-  from: MessageEventFrom
-  type: MessageEventTypes
-  typing: number
-  delay: number
-}
+export * from './buttons'
+export * from './carousel'
+export * from './custom'
+export * from './location'
+export * from './media'
+export * from './message-event'
+export * from './postback'
+export * from './replies'
+export * from './text'
