+++ packages/botonic-core/src/models/events/message/location.ts
@@ -1,4 +1,4 @@
-import { BotonicMessageEvent, MessageEventTypes } from '.'
+import { BotonicMessageEvent, MessageEventTypes } from './message-event'
 
 export interface LocationMessageEvent extends BotonicMessageEvent {
   type: MessageEventTypes.LOCATION
