+++ packages/botonic-core/src/models/events/message/postback.ts
@@ -1,4 +1,4 @@
-import { BotonicMessageEvent } from '.'
+import { BotonicMessageEvent } from './message-event'
 
 export interface PostbackMessageEvent extends BotonicMessageEvent {
   payload: string
