+++ packages/botonic-core/src/models/events/message/text.ts
@@ -1,5 +1,5 @@
-import { BotonicMessageEvent, MessageEventTypes } from '.'
 import { WithButtons } from './buttons'
+import { BotonicMessageEvent, MessageEventTypes } from './message-event'
 import { WithReplies } from './replies'
 
 export interface TextMessageEvent
