+++ packages/botonic-core/src/models/index.ts
@@ -1,0 +1,3 @@
+export * from './events'
+export * from './legacy-types'
+export * from './user'
